# src/docs/openapi.yaml
openapi: "3.0.0"
info:
  title: Shopix E-Commerce API
  version: "1.0.0"
  description: |
    Full REST API documentation for the **Shopix** e-commerce platform.

    ### Authentication
    Most endpoints require a valid `Accesstoken` cookie set after login.
    Pass credentials via the login endpoint — the server sets HttpOnly cookies automatically.

  contact:
    name: Shopix Support
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local Development
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: Accesstoken
  schemas:
...
    Store:
      type: object
      properties:
        _id:
          type: string
        sellerId:
          type: string
        name:
          type: string
        description:
          type: string
        type:
          type: string
        logo:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected, frozen]
        createdAt:
          type: string
          format: date-time

    # ─── Conversation ─────────────────────────────────────────
...
    Review:
      type: object
      properties:
        _id:
          type: string
        productId:
          type: string
        userId:
          type: object
          properties:
            _id:
              type: string
            name:
              type: string
        rating:
          type: number
          minimum: 1
          maximum: 5
        comment:
          type: string
        createdAt:
          type: string
          format: date-time

# ═══════════════════════════════════════════════════════════════
# PATHS
# ═══════════════════════════════════════════════════════════════
paths:
...
  /api/reviews:
    post:
      tags: [Reviews]
      summary: Submit a product review
      description: Creates a review for a product and recalculates the product's average rating.
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [productId, rating, comment]
              properties:
                productId:
                  type: string
                rating:
                  type: number
                  minimum: 1
                  maximum: 5
                  example: 4
                comment:
                  type: string
                  example: Great product!
      responses:
        "200":
          description: Review submitted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  review:
                    $ref: "#/components/schemas/Review"
        "400":
          description: All fields required

  # ───────────────────────────────────────────────────────────
  # STORE (Seller Dashboard)
  # ───────────────────────────────────────────────────────────
...
  /api/store/admin/review:
    get:
      tags: [Admin – Stores]
      summary: List all stores for review (admin)
      description: Returns all stores with seller info, ordered by newest first. Admin only.
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Stores list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Store"
        "403":
          description: Unauthorized — admin only

    put:
      tags: [Admin – Stores]
      summary: Approve / reject / freeze a store (admin)
      description: Updates the store status. Sends corresponding email and in-app notification to the seller.
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [storeId, status]
              properties:
                storeId:
                  type: string
                status:
                  type: string
                  enum: [approved, rejected, frozen]
      responses:
        "200":
          description: Store status updated
        "400":
          description: storeId and status required
        "403":
          description: Unauthorized — admin only
        "404":
          description: Store not found

    delete:
      tags: [Admin – Stores]
      summary: Delete a store (admin)
      description: Permanently deletes a store and its Cloudinary logo after admin password verification.
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [storeId, password]
              properties:
                storeId:
                  type: string
                password:
                  type: string
                  format: password
      responses:
        "200":
          description: Store deleted
        "401":
          description: Incorrect admin password
        "403":
          description: Unauthorized — admin only
        "404":
          description: Store not found
...